WIP:hooks.c: Allow host/service check propagation from passive check results
When receiving an external command over merlin, we set merlin_sender to the sending node.
If the external command is a passive check result and we are the responsible node
we should generate a check result and propagate it to our fellow nodes. However
we would not be able to send the check result because we'd set the packet to MAGIC_NONET
because merlin_sender was set.

Check results should be set to MAGIC_NONET if merlin_sender is set IF merlin_sender is
set as a result of receiving a check result over merlin and NOT an external command.

Signed-off-by: Robin Engström <robin.engstrom@op5.com>